requirements.txt
numpy
pandas
scikit-learn
librosa
soundfile
joblib
tqdm
matplotlib

# Optional for ASR
openai-whisper
transformers
torchaudio
